Graphs and algorithm fix.

+++ src/main/java/com/josephaines/sga/Data.java
@@ -1,13 +1,13 @@
 package com.josephaines.sga;
 
 public class Data {
-    public float[] averageUtility;
-    public float[] maxUtility;
-    public float[] minUtility;
+    public double[] averageUtility;
+    public double[] maxUtility;
+    public double[] minUtility;
 
     public Data(int cycles) {
-        averageUtility = new float[cycles];
-        maxUtility = new float[cycles];
-        minUtility = new float[cycles];
+        averageUtility = new double[cycles];
+        maxUtility = new double[cycles];
+        minUtility = new double[cycles];
     }
 }
